refactor: `AddCardUiState`를 `addCard` 패키지로 이동
